Update: Dependencies working & canvas w/objects
Got the dependencies working best with using a VM option by editing the configuration for the runtime. In the VM option you can use (--module-path "Path to javafx sdk" --add-modules javafx.controls,javafx.fxml,javafx.media,javafx.web)  this one is for windows but there are mac and linux commands online you can find or you can translate it. I added .idea to the .gitIgnore so I'm not pushing a bunch of unnecessary things either. I used the ball that was drawn in by Jacob and created a ground for the ball to land on with the gravity factor and changed the package names for anything under "physics" because my package was named "Physics2D".  Lastly I manually ported all the files created on the main branch to the shapes_and_basics branch and commented out the unused code I'd written that was causing errors. I will go through and delete unnecessary code on my next commit, in the future I'll stage commits in smaller chunks so that version control is better and more manageable.